# src/router/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin Portal API",
    "description": "Rest API for Admin Portal",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logs the user out",
        "description": "Deletes the session from the database",
        "parameters": [
          {
            "name": "sessionToken",
            "description": "Token generated by the server",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}